requirements.txt
asgiref==3.2.7
Django==3.0.6
django-crispy-forms==1.9.1
django-social-share==1.4.0
django-taggit==1.3.0
importlib-metadata==1.6.0
Markdown==3.2.2
pep8==1.7.1
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
